move namespace_from_domstring to script::dom::bindings::domname
